Safari-609.1.11.1/LayoutTests/inspector/canvas/console-record-webgl2-expected.txt

Test that CanvasManager is able to record actions made to WebGL2 canvas contexts.


== Running test suite: Canvas.recordingWebGL2
-- Running test case: Canvas.recordingWebGL2.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
